transform the view implementation into a component called SearchComponent
